+++ edsnlp/pipelines/core/normalizer/normalizer.py
@@ -6,17 +6,19 @@ from .accents import Accents
 from .lowercase import remove_lowercase
 from .pollution import Pollution
 from .quotes import Quotes
+from .spaces import Spaces
 
 
 class Normalizer(object):
     """
     Normalisation pipeline. Modifies the `NORM` attribute,
-    acting on four dimensions :
+    acting on five dimensions :
 
     - `lowercase`: using the default `NORM`
     - `accents`: deterministic and fixed-length normalisation of accents.
     - `quotes`: deterministic and fixed-length normalisation of quotation marks.
-    - `pollution`: removal of pollutions.
+    - `spaces`: "removal" of spaces tokens (via the tag_ attribute).
+    - `pollution`: "removal" of pollutions (via the tag_ attribute).
 
     Parameters
     ----------
@@ -26,6 +28,8 @@ class Normalizer(object):
         Optional `Accents` object.
     quotes : Optional[Quotes]
         Optional `Quotes` object.
+    spaces : Optional[Spaces]
+        Optional `Spaces` object.
     pollution : Optional[Pollution]
         Optional `Pollution` object.
     """
@@ -35,11 +39,13 @@ class Normalizer(object):
         lowercase: bool,
         accents: Optional[Accents],
         quotes: Optional[Quotes],
+        spaces: Optional[Spaces],
         pollution: Optional[Pollution],
     ):
         self.lowercase = lowercase
         self.accents = accents
         self.quotes = quotes
+        self.spaces = spaces
         self.pollution = pollution
 
     def __call__(self, doc: Doc) -> Doc:
@@ -62,7 +68,8 @@ class Normalizer(object):
             self.accents(doc)
         if self.quotes is not None:
             self.quotes(doc)
+        if self.spaces is not None:
+            self.spaces(doc)
         if self.pollution is not None:
             self.pollution(doc)
-
         return doc