+++ edsnlp/pipelines/ner/drugs/factory.py
@@ -9,6 +9,7 @@ from . import patterns
 DEFAULT_CONFIG = dict(
     attr="NORM",
     ignore_excluded=False,
+    ignore_space_tokens=False,
     term_matcher=TerminologyTermMatcher.exact,
     term_matcher_config={},
 )
@@ -21,12 +22,40 @@ DEFAULT_CONFIG = dict(
 )
 def create_component(
     nlp: Language,
-    name: str,
-    attr: str,
-    ignore_excluded: bool,
-    term_matcher: TerminologyTermMatcher,
-    term_matcher_config: Dict[str, Any],
+    name: str = "eds.drugs",
+    attr: str = "NORM",
+    ignore_excluded: bool = False,
+    ignore_space_tokens: bool = False,
+    term_matcher: TerminologyTermMatcher = TerminologyTermMatcher.exact,
+    term_matcher_config: Dict[str, Any] = {},
 ):
+    """
+    Create a new component to recognize and normalize drugs in documents.
+    The terminology is based on Romedi (see documentation) and the
+    drugs are normalized to the ATC codes.
+
+    Parameters
+    ----------
+    nlp: Language
+        spaCy `Language` object.
+    name: str
+        The name of the pipe
+    attr: Union[str, Dict[str, str]]
+        Attribute to match on, eg `TEXT`, `NORM`, etc.
+    ignore_excluded: bool
+        Whether to skip excluded tokens during matching.
+    ignore_space_tokens: bool
+        Whether to skip space tokens during matching.
+    term_matcher: TerminologyTermMatcher
+        The term matcher to use, either `TerminologyTermMatcher.exact` or
+        `TerminologyTermMatcher.simstring`
+    term_matcher_config: Dict[str, Any]
+        The configuration for the term matcher
+
+    Returns
+    -------
+    TerminologyMatcher
+    """
     return TerminologyMatcher(
         nlp,
         label="drug",
@@ -34,6 +63,7 @@ def create_component(
         regex=dict(),
         attr=attr,
         ignore_excluded=ignore_excluded,
+        ignore_space_tokens=ignore_space_tokens,
         term_matcher=term_matcher,
         term_matcher_config=term_matcher_config,
     )